+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 1.1.0 - 2019-11-17
+### Added
+ - Clips popout support
+ - Apply volume setting
+ - Apply window position setting
+ - Player controls feature
+ - Community points auto claim bonus feature
+ - Invisible space at the end of timestamp
+ - Option to prevent chat badge selection
+
+### Changed
+ - Options page organization
+
+### Fix
+ - Options page now reset properly
+
 ## 1.0.1 - 2019-09-29
 ### Changed
  - Chat header link adder updated to new Twitch style